requirements.txt
python-nmap
python-whois
colorama
ollama